implement user status model

+++ app/Http/Controllers/Admin/AdminController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Comment;
 use App\Models\User;
 use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Auth;
 
 
 class AdminController extends Controller
@@ -39,12 +40,18 @@ class AdminController extends Controller
 
     public function deleteUser($id)
     {
-        User::destroy($id);
+        if(Auth::id() != $id){
+            User::destroy($id);
+        }else{
+            notice("You can't delete yourself", 'warning');
+            return redirect()->back();
+        }
 
         if(!is_null(User::find($id))){
             notice("Something went wrong", 'warning');
             return redirect()->back();
         }
+
         notice("User deleted", 'success');
         return redirect()->back();
     }
@@ -63,21 +70,37 @@ class AdminController extends Controller
 
     public function blockUser($id)
     {
+        if(Auth::id() != $id){
+            User::findOrFail($id)->setBanned();
+        }else{
+            notice("You can't block yourself", 'warning');
+            return redirect()->back();
+        }
 
+        notice("User blocked", 'success');
+        return redirect()->back();
     }
 
     public function unblockUser($id)
     {
+        User::findOrFail($id)->removeBanned();
 
+        notice("User unblocked", 'success');
+        return redirect()->back();
     }
+
     public function activateUser($id)
     {
+        User::findOrFail($id)->setActive();
 
-
+        notice("User activated", 'success');
+        return redirect()->back();
     }
     public function deactivateUser($id)
     {
+        User::findOrFail($id)->setWait();
 
-
+        notice("User deactivated", 'success');
+        return redirect()->back();
     }
 }
